Enum ddsfile::D3DFormat

pub enum D3DFormat {

    A8B8G8R8,
    G16R16,
    A2B10G10R10,
    A1R5G5B5,
    R5G6B5,
    A8,
    A8R8G8B8,
    X8R8G8B8,
    X8B8G8R8,
    A2R10G10B10,
    R8G8B8,
    X1R5G5B5,
    A4R4G4B4,
    X4R4G4B4,
    A8R3G3B2,
    A8L8,
    L16,
    L8,
    A4L4,
    DXT1,
    DXT3,
    DXT5,
    R8G8_B8G8,
    G8R8_G8B8,
    A16B16G16R16,
    Q16W16V16U16,
    R16F,
    G16R16F,
    A16B16G16R16F,
    R32F,
    G32R32F,
    A32B32G32R32F,
    DXT2,
    DXT4,
    UYVY,
    YUY2,
    CXV8U8,

}

Variants

A8B8G8R8

G16R16

A2B10G10R10

A1R5G5B5

R5G6B5

A8

A8R8G8B8

X8R8G8B8

X8B8G8R8

A2R10G10B10

R8G8B8

X1R5G5B5

A4R4G4B4

X4R4G4B4

A8R3G3B2

A8L8

L16

L8

A4L4

DXT1

DXT3

DXT5

R8G8_B8G8

G8R8_G8B8

A16B16G16R16

Q16W16V16U16

R16F

G16R16F

A16B16G16R16F

R32F

G32R32F

A32B32G32R32F

DXT2

DXT4

UYVY

YUY2

CXV8U8

Implementations

impl D3DFormat

pub fn r_bit_mask(&self) -> Option<u32>

This gets the bitmask for the red channel pixels

pub fn g_bit_mask(&self) -> Option<u32>

This gets the bitmask for the green channel pixels

pub fn b_bit_mask(&self) -> Option<u32>

This gets the bitmask for the blue channel pixels

pub fn a_bit_mask(&self) -> Option<u32>

This gets the bitmask for the alpha channel pixels

pub fn try_from_pixel_format(pixel_format: &PixelFormat) -> Option<D3DFormat>

This attempts to use PixelFormat data (e.g. from the dds.header.spf field) to determine the D3DFormat.